Revert "Backup segregation"
