+++ packages/core/src/implementations/data/AdDataRepository.ts
@@ -5,6 +5,7 @@ import {
   AdSignature,
   VolatileStorageMetadata,
 } from "@snickerdoodlelabs/objects";
+import { ERecordKey } from "@snickerdoodlelabs/persistence";
 import { inject, injectable } from "inversify";
 import { ResultAsync } from "neverthrow";
 import { ResultUtils } from "neverthrow-result-utils";
@@ -14,7 +15,6 @@ import {
   IDataWalletPersistence,
   IDataWalletPersistenceType,
 } from "@core/interfaces/data/index.js";
-import { ERecordKey } from "@snickerdoodlelabs/objects";
 
 @injectable()
 export class AdDataRepository implements IAdDataRepository {
