+++ packages/core/src/implementations/data/BrowsingDataRepository.ts
@@ -7,6 +7,7 @@ import {
   ClickData,
   VolatileStorageMetadata,
 } from "@snickerdoodlelabs/objects";
+import { ERecordKey } from "@snickerdoodlelabs/persistence";
 import { inject, injectable } from "inversify";
 import { ResultAsync, okAsync } from "neverthrow";
 import { ResultUtils } from "neverthrow-result-utils";
@@ -17,7 +18,6 @@ import {
   IDataWalletPersistence,
   IDataWalletPersistenceType,
 } from "@core/interfaces/data/index.js";
-import { ERecordKey } from "@snickerdoodlelabs/objects";
 
 @injectable()
 export class BrowsingDataRepository implements IBrowsingDataRepository {
