+++ packages/core/src/implementations/data/DemographicDataRepository.ts
@@ -9,6 +9,7 @@ import {
   EmailAddressString,
   CountryCode,
 } from "@snickerdoodlelabs/objects";
+import { EFieldKey } from "@snickerdoodlelabs/persistence";
 import { inject, injectable } from "inversify";
 import { ResultAsync } from "neverthrow";
 
@@ -17,7 +18,6 @@ import {
   IDataWalletPersistenceType,
   IDemographicDataRepository,
 } from "@core/interfaces/data/index.js";
-import { EFieldKey } from "@snickerdoodlelabs/objects";
 
 @injectable()
 export class DemographicDataRepository implements IDemographicDataRepository {
