+++ packages/core/src/implementations/data/LinkedAccountRepository.ts
@@ -14,9 +14,8 @@ import {
   Signature,
   TokenId,
   VolatileStorageMetadata,
-  ERecordKey,
-  EFieldKey,
 } from "@snickerdoodlelabs/objects";
+import { EFieldKey, ERecordKey } from "@snickerdoodlelabs/persistence";
 import { inject, injectable } from "inversify";
 import { ResultAsync } from "neverthrow";
 import { ResultUtils } from "neverthrow-result-utils";
