+++ packages/objects/src/enum/index.ts
@@ -1,4 +1,3 @@
-export * from "@objects/enum/ELocalStorageKey";
 export * from "@objects/enum/EAdContentType.js";
 export * from "@objects/enum/EBackupPriority.js";
 export * from "@objects/enum/EChain.js";
