+++ packages/persistence/src/backup/index.ts
@@ -1,6 +1,4 @@
 export * from "@persistence/backup/BackupManager.js";
 export * from "@persistence/backup/BackupManagerProvider.js";
 export * from "@persistence/backup/IBackupManager.js";
-export * from "@persistence/backup/ChunkRenderer.js";
-export * from "@persistence/backup/IChunkRenderer.js";
 export * from "@persistence/backup/IBackupManagerProvider.js";
