+++ packages/persistence/src/cloud/GoogleCloudStorage.ts
@@ -110,7 +110,6 @@ export class GoogleCloudStorage implements ICloudStorage {
       .andThen(([privateKey, config]) => {
         const addr =
           this._cryptoUtils.getEthereumAccountAddressFromPrivateKey(privateKey);
-
         return this.insightPlatformRepo.clearAllBackups(
           privateKey,
           config.defaultInsightPlatformBaseUrl,
