+++ packages/test-harness/src/prompts/CorePrompt.ts
@@ -322,9 +322,6 @@ export class CorePrompt extends DataWalletPrompt {
               const backupChoices = chunks.map((chunk) => {
                 return new BackupChoice(chunk);
               });
-              if (backupChoices.length == 0) {
-                return okAsync(null);
-              }
               return inquiryWrapper({
                 type: "list",
                 name: "backupPrompt",
@@ -333,9 +330,6 @@ export class CorePrompt extends DataWalletPrompt {
               });
             })
             .andThen((selection) => {
-              if (selection == null){
-                return okAsync(console.log("No backups found"));
-              }
               return this.core
                 .fetchBackup(selection.backupPrompt)
                 .andThen((output) => {
