+++ packages/test-harness/src/utilities/DataWalletProfile.ts
@@ -417,7 +417,6 @@ export class DataWalletProfile {
             timestamp: UnixTimestamp(backupJson.timestamp),
             signature: backupJson.signature,
             priority: EBackupPriority.NORMAL,
-            dataType: backupJson,
           },
           blob: new AESEncryptedString(
             EncryptedString(backupJson.blob.data),
